+++ apps/xim2_web/lib/xim2_web/router.ex
@@ -26,6 +26,7 @@ defmodule Xim2Web.Router do
     live "/components", ComponentLive.Index, :index
     live "/astrorunner", AstrorunnerLive.Index, :index
     live "/my_liege", MyLiegeLive.Index, :index
+    live "/my_liege/scenario", MyLiegeLive.Scenario, :index
   end
 
   # Other scopes may use custom stacks.
