+++ apps/xim2_web/test/xim2_web/live/my_liege_live/scenario_test.exs
@@ -1,4 +1,4 @@
-defmodule Xim2Web.MyLiegeLive.IndexTest do
+defmodule Xim2Web.MyLiegeLive.ScenarioTest do
   use Xim2Web.ConnCase
   import Phoenix.LiveViewTest
 
@@ -11,12 +11,12 @@ defmodule Xim2Web.MyLiegeLive.IndexTest do
   end
 
   test "index", %{conn: conn} do
-    {:ok, view, _html} = live(conn, "/my_liege")
+    {:ok, view, _html} = live(conn, "/my_liege/scenario")
     assert has_element?(view, "h1", "My Liege")
   end
 
   test "change food", %{conn: conn} do
-    {:ok, view, _html} = live(conn, "/my_liege")
+    {:ok, view, _html} = live(conn, "/my_liege/scenario")
     assert view |> element("#property-storage-food") |> render() =~ "94"
     assert view |> element("#property-storage-food a") |> render_click() =~ "94"
 
@@ -28,14 +28,14 @@ defmodule Xim2Web.MyLiegeLive.IndexTest do
   end
 
   test "sim step", %{conn: conn} do
-    {:ok, view, _html} = live(conn, "/my_liege")
+    {:ok, view, _html} = live(conn, "/my_liege/scenario")
     assert view |> element("#social_stratum_poverty .gen_1") |> render() =~ "10"
     view |> element("#button-sim-step") |> render_click()
     refute view |> element("#social_stratum_poverty .gen_1") |> render() =~ "10"
   end
 
   test "recreate", %{conn: conn} do
-    {:ok, view, _html} = live(conn, "/my_liege")
+    {:ok, view, _html} = live(conn, "/my_liege/scenario")
     assert view |> element("#social_stratum_poverty .gen_1") |> render() =~ "10"
     view |> element("#button-sim-step") |> render_click()
     refute view |> element("#social_stratum_poverty .gen_1") |> render() =~ "10"
